Add feature to add PDO mapping to SII file
For devices which support PDO assignment or PDO configuration the use of a
fixed PDO configuration is neither not recommendet or completete senseless.
If present this could confuse the master and lead to unexpected behaviour.

With this patch the commandline parameter `-m` is added to if the PDO mapping
should be part of the SII file.  This option does not affect the pretty print
parameter `-p`.