Fix potential endless loop
If the end of file marker is missing the content parser can be caught in
a endless loop and never abort. This patch will abort after all bytes of
the input are consumed and print an error if no 0xffff (end of file) is
found. With the `-p` a best efford to print everything which is deteced
is done.