Add PDO flag handling
Fixes: https://www.wrike.com/open.htm?id=967735250

* Add parsing of flags for PDO

In ETG 2000 v1.0.14 only the flags

- Fixed
- Mandatory
- Virtual
- OverwrittenByModule

This can be expanded with more if necessary.

* Parsing of PDO entry flag

The entry flags are no longer printed. In ETG2010 (v1.0.0) these flags
are marked as 'for future use'. In ETG2000 only the "fixed" flag is
mentioned but there is no definit position to put it. So I leave it out
for now.

However, the ESI description can contain a fixed flag so I add it to the
internal data structure in order to use it as soon as ETG2010 describes
where to put it.

* Add True value parsing

This will read ['True', 'true', '1'] as boolean true value. If this is
not found then it is false.

First use is the parsing of the [RT]xPDO flags.