Change DC structure
Fixed the DC sructure according to ETG 2010 (v1.0.0). It looks like the
value definition is not complete. For example I would expect something
like CycleTimeSync1 (or am I mistaken here?).

Remove unnecessary cast step:

The parse_*_section() functions return the correct struct type but the
data field in the _sii_cat struct does not really bother about the
specific type. The knowledge about the struct of type of the section
does not really matter in this context and does not add any clearity.

Refactor parse_dclock() function to read all opmodes:

The function parse_dclock() had the error that it assumed there is only
one Dc opmode. I added a outer loop to iterate over all the OpMode in
the Dc section of the ESI file. It is possible that there are multiple
OpModes for on device and now the parser addresses this fact. The SII
reading already worked with multiple DC configurations. Furthermore the
values written in the ESI are now taken into consideration.

Add cycleTime1 parameter to the DC struct to configure SYNC1 times.

This resolves issue #16.